Rename new test suites
- keep legacy tests in the `tests.v0` package
- prefix legacy test suites with "Legacy"
