+++ scalafix-tests/shared/src/main/scala/test/LegacySyntheticsTest.scala
@@ -1,8 +1,10 @@
+package test
+
 import scala.collection.breakOut
 import scala.concurrent.Future
 import scala.concurrent.ExecutionContext.Implicits.global
 
-class SyntheticTest(xs: List[Int]) {
+class LegacySyntheticsTest(xs: List[Int]) {
   def m1(xs: Set[Int]): List[Int] =
     xs.to
